fixed enemy AI
Enemies now spawn sporadically around the player with a limit of 50, their ai involves moving to a random position if they do not see the player, and chasing the player until they leave a certain radius if they do see the player. Their vision is cone-based with three rays in their front. basically every parameter is adjustable